Change both mkCharToBePut and traverseMaskChars objects
A prop named mode was added to the traverseMaskChars' state object. This
prop aim to make masking and unmasking available in the same fn.

A prop named regExp was added to the object returned by mkCharToBePut,
which is a prop that others characters already have in their returned
object. The need for this prop to be present in the mkCharToBePut's
object appeared when implementing the unmask mode in the
traverseMaskChars fn, since, when unmasking, it needs to be checked
whether a maskedValue's character matches a maskChars' toBePut.